requirements.txt
Flask
Flask-SocketIO
python-dotenv
requests
gunicorn
gevent
gevent-websocket
psycopg2-binary
langchain
langchain-community
jinja2
python-dateutil